Added full functionality for scrolling with mouse, nav buttons, and arrow keys
