+++ app/src/index.css
@@ -1,6 +1,6 @@
 @import "tailwindcss";
 
-:root {
+@theme {
     --breakpoint-md: 600px;
 }
